serverless/landmark-recognition/detect-task/requirements.txt
google-auth==1.0.2
google-cloud-vision==0.27.0
requests==2.18.4
fdk
